Added comment to explain why you get "Failed to restore data: #{self.filename}" notice after running `node server.js` locally for the first time.
The message previously made me think I had something wrong with my file/directory permissions on `/tmp`. Permissions were fine - and I eventually realised the true cause of the "failure" was an empty file - which obviously can not be restored from.

The note I've added should help to ease the minds of others running the server for the first time... but ideally a better solution should be looked into (i.e. don't report the failure if the file was read successfully, but was just empty).